refactor: ValidationCheck.js 생성 및 관련 메서드 모듈화
App에서 담당하는 여러기능의 역할을 나누기 위해 사용자 숫자 입력 유효성 검사 메서드를 분리 하였음.
